Changed file structure, moved profile preview into its own component
