+++ .env
@@ -1,0 +1,1 @@
+WATCHPACK_POLLING=true
